kanui/requirements.txt
scrapy==2.5.0
scrapy-magicfields==1.1.0
google-cloud-storage==1.31.2
google-auth==1.23.0
discord-webhook==0.14.0
typer==0.4.0
dynaconf==3.1.5